fix(hir): #6233 review follow-ups — lint file cap + non-class shadow typing
- lint: lower_types.rs hit 2007 lines (2000-line file cap) after the
  #6233 guards. Split per the repo recipe: the TS-annotation extraction
  group (extract_ts_type & friends, decorator lowering) moves to
  lower_types/extract.rs as a pure re-exported submodule — named
  re-exports keep every existing call path compiling unchanged. The six
  `use crate::lower_types::*` globs that became redundant are removed
  or replaced with named imports (back to the exact pre-change warning
  count).
- CodeRabbit Major: the type-inference shadow guards only honored user
  CLASSES — a local, `function Map() {}`, or import shadowing a
  built-in name still got the built-in declared type, routing method
  calls on the constructed instance down the intrinsic fast paths.
  Both inference sites (infer_type_from_expr's New arm +
  url_encoding_constructor_type, and the var-decl chain in
  destructuring/var_decl/type_infer.rs) now mirror the
  construction-side predicate (shadows_unqualified_global), scoped to
  a conservative builtin_constructor_inference_name set so
  module-export names that legitimately arrive through local bindings
  (Buffer, the stream and event classes) keep their arms.
- CodeRabbit Minor (outside-diff): `new Map<number>()` on a generic
  user `class Map<T>` took the explicit-type-args early return and
  still produced `Generic { base: "Map" }` — exactly what the
  collection fast-path recognizers key on. The user-class check is
  hoisted above that early return; built-in-colliding user classes
  stay `Named` even with explicit type args (both inference sites).
- New regression test test_issue_6233_shadowed_global_fn_new.ts:
  function/const constructors named Map / Set / Uint8Array /
  URLSearchParams whose instances carry intrinsic-colliding members
  (get/set/has/size/length) — matches node byte-for-byte.
