+++ crates/perry-hir/src/destructuring/var_decl/alias_tracking.rs
@@ -10,7 +10,6 @@ use swc_ecma_ast as ast;
 use crate::ir::*;
 use crate::lower::{lower_expr, LoweringContext};
 use crate::lower_patterns::*;
-use crate::lower_types::*;
 
 use crate::destructuring::var_decl_sources::*;
